Export get_random_avatar to account utils
