+++ saleor/graphql/account/mutations.py
@@ -1,12 +1,9 @@
-import os
-import random
 from textwrap import dedent
 
 import graphene
 from django.conf import settings
 from django.contrib.auth.tokens import default_token_generator
 from django.core.exceptions import ObjectDoesNotExist
-from django.core.files import File
 from django.utils.encoding import force_bytes
 from django.utils.http import urlsafe_base64_encode
 from graphql_jwt.decorators import (
@@ -15,6 +12,7 @@ from graphql_jwt.exceptions import PermissionDenied
 
 from ...account import emails, models, utils
 from ...account.thumbnails import create_user_avatar_thumbnails
+from ...account.utils import get_random_avatar
 from ...checkout import AddressType
 from ...core.permissions import MODELS_PERMISSIONS, get_permissions
 from ...dashboard.emails import (
@@ -258,17 +256,10 @@ class StaffCreate(ModelMutation):
         return cleaned_input
 
     @classmethod
-    def get_avatar(cls):
-        avatars_dir = 'saleor/static/images/avatars/'
-        avatar_name = random.choice(os.listdir(avatars_dir))
-        avatar_path = os.path.join(avatars_dir, avatar_name)
-        return File(open(avatar_path, 'rb'), name=avatar_name)
-
-    @classmethod
     def save(cls, info, user, cleaned_input):
-        user.avatar = cls.get_avatar()
+        user.avatar = get_random_avatar()
         user.save()
-        create_user_avatar_thumbnails(user_id=user.pk)
+        create_user_avatar_thumbnails.delay(user_id=user.pk)
         if cleaned_input.get('send_password_email'):
             send_set_password_staff_email.delay(user.pk)
